coerce csv,tsv to officedocs
This is because Google Drive fails to convert text/csv types
to Spreadsheets, however with the Spreadsheet coercion it alludes
them to Spreadsheets but on final open it will fully convert it to a
Google Sheet. See the tricky thing though is that trying to force
csv/tsv directly to Google Sheets fails since Google Docs catches this
and then reverts them to text documents.