unknown errors now retryable by default
In relation to a large set of unknown errors, moreover looking at
https://github.com/google/google-api-go-client/issues/93
where we are type casting to a googleapi.Error yet errors could be
dial/net related errors, it is safer to make the largely unknown errors
retryable.

This actually helps a lot say when downloading files and the connection
is cut out maybe due to WiFi trips or going underground.