pull: export exportables on mod without checksum
Fixes #496.

Now allows for docs to be exported when modified since before
there was a clause that prevented files from being downloaded
if their checksums varied, of which exportable docs have no
raw content hence their checksums will always be the same
yet this is registered as only a ModTime change.