pull/export: fixed ppt vs pptx regex ambiguity
Fixes #583.
Updates #551.

Fixes an ambiguity with the regexes for `ppt` and `pptx`.
The problem was that `ppt` and `pptx` are unterminated regex-wise
so `pptx` would get matched as `ppt` and then undeterministically
also matched as `pptx`.
This mean that `pptx` would give two different mimeTypes ie
```json
	"ppt":    "application/vnd.ms-powerpoint",
	"pptx":   "application/vnd.openxmlformats-officedocument.presentationml.presentation"
```

The fix is to terminate them as `ppt$` and `pptx$` to properly
differentiate between the two.