push/mounted: fix unload of paths in root + no return catch
Fixes https://github.com/odeke-em/drive/issues/641.

+ Refactored push mounted, removing/generalizing code for
mounts to use same as ordinary pushes.

+ Ensure that in cmd/drive/main.go, after a mounted push
we return appropriately otherwise we were erraneously falling
to start an ordinary push which would trigger a bug with
a bad context. This is akin to the common C-style switch bug in which
one forgets to break/return.

+ Fixed up a bug with mounted push where if the user
specified a single mounted path, the current working
directory wouldn't be added to the arguments that needed
to be processed. It was an error comparison mismatch.