handle differing dirTypes on direct addressing
Fixes https://github.com/odeke-em/drive/issues/769.

Handle the case where the local and remote objects
of a directly addressed path have different dirTypes ie
one is a regular file, the other is a directory/folder.
It returns an IllogicalStateError whose status can
be checked in the shell.

Fixes a TODO that has been waiting to be fixed.
Ensures that the case is caught e.g

* Exhibit:
```shell
$ drive new --mime-key docs share_test
$ mkdir -p share_test
$ drive pull share_test; echo $?
Resolving...
/share_test: local is a directory while remote is a file
8
```