stat/md5: allow op on depth=0 for dir children
Fixes https://github.com/odeke-em/drive/issues/783.

Ensures that for a directory, after decrementing the depth
by 1 in the parent's first pass that we at least operate on
passed in files instead of erraneously ending on that depth.
Traversal depth is meant for traversing directories/folders
and less on actual files.

This change leads us to conform to the proper behavior exhibited
by  `drive list`.